@@ -31,19 +31,21 @@
     ]
   },
   "dependencies": {
-    "@date-io/date-fns": "2.6.1",
+    "@date-io/date-fns": "^1.3.13",
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
     "@formatjs/intl-pluralrules": "1.3.5",
     "@fortawesome/fontawesome-free": "5.13.0",
     "@manaflair/redux-batch": "1.0.0",
-    "@material-ui/core": "4.9.14",
-    "@material-ui/icons": "4.9.1",
-    "@material-ui/lab": "4.0.0-alpha.53",
-    "@material-ui/pickers": "3.2.10",
-    "@material-ui/styles": "4.9.14",
-    "@reduxjs/toolkit": "1.3.6",
+    "@mui/material": "^5.1.0",
+    "@mui/icons-material": "^5.1.0",
+    "@mui/lab": "^5.0.0-alpha.54",
+    "@mui/x-date-pickers": "^5.0.11",
+    "@mui/styles": "^5.1.0",
+    "@reduxjs/toolkit": "^1.6.0",
     "@tanem/svg-injector": "8.0.50",
-    "apexcharts": "3.19.2",
-    "axios": "0.19.2",
+    "apexcharts": "^3.27.1",
+    "axios": "^0.21.1",
     "axios-mock-adapter": "1.18.1",
     "bootstrap": "4.5.0",
     "clipboard-copy": "3.1.0",
@@ -58,15 +60,17 @@
     "jss-rtl": "^0.3.0",
     "lodash": "4.17.15",
     "material-ui-popup-state": "1.4.1",
+    "moment": "^2.29.1",
+    "moment-timezone": "^0.5.34",
     "object-path": "0.11.4",
     "perfect-scrollbar": "1.5.0",
     "prop-types": "15.7.2",
-    "react": "16.12.0",
+    "react": "^17.0.2",
     "react-bootstrap": "1.0.1",
     "react-bootstrap-table-next": "4.0.2",
     "react-bootstrap-table2-paginator": "2.1.2",
     "react-datepicker": "2.16.0",
-    "react-dom": "16.12.0",
+    "react-dom": "^17.0.2",
     "react-draggable": "4.4.2",
     "react-inlinesvg": "1.2.0",
     "react-intl": "3.6.2",
