@@ -1,5 +1,5 @@
 import React from "react";
-import { CircularProgress } from "@material-ui/core";
+import { CircularProgress } from "@mui/material";
 import { toAbsoluteUrl } from "../../_helpers";
 
 export function SplashScreen() {
