@@ -7,7 +7,7 @@ import {
   Divider,
   Fade,
   Popper,
-} from "@material-ui/core";
+} from "@mui/material";
 import React, { useState } from "react";
 import { useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
