@@ -27,6 +27,6 @@ export * from "./Card";
 export { default as AnimateLoading } from "./AnimateLoading";
 export { LoadingDialog } from "./LoadingDialog";
 export { Notice } from "./Notice";
-export { SplashScreen } from "./SplashScreen";
+// export { SplashScreen } from "./SplashScreen";
 export { SVGIcon } from "./SVGIcon";
 export { ModalProgressBar } from "./ModalProgressBar";
