@@ -4,6 +4,7 @@ import objectPath from "object-path";
 import { Brand } from "../brand/Brand";
 import { AsideMenu } from "./aside-menu/AsideMenu";
 import { useHtmlClassService } from "../../_core/MetronicLayout";
+import asideBg from "../../../../assets/aside-bg.jpg";
 
 export function Aside() {
   const uiService = useHtmlClassService();
@@ -27,8 +28,6 @@ export function Aside() {
         style={{ backgroundColor: "#1a1a27" }}
         id="kt_aside"
         className={`aside aside-left  ${layoutProps.asideClassesFromConfig} d-flex flex-column flex-row-auto`}
-        // id="kt_aside"
-        // className={`aside aside-left  ${layoutProps.asideClassesFromConfig} d-flex flex-column flex-row-auto`}
       >
         <Brand />
 
@@ -37,12 +36,13 @@ export function Aside() {
           id="kt_aside_menu_wrapper"
           className="aside-menu-wrapper flex-column-fluid"
           style={{
-            // backgroundImage: `url(${asideBg})`,
-            // backgroundPosition: "center",
-            flex: 1,
-            minHeight: "0px",
-            display: "flex",
-            flexDirection: "column",
+           backgroundImage: `url(${asideBg})`,
+           backgroundRepeat: "no-repeat",
+           backgroundPosition: "center center",
+           flex: 1,
+           minHeight: "0px",
+           display: "flex",
+           flexDirection: "column",
           }}
         >
           {layoutProps.disableAsideSelfDisplay && (
