@@ -1,11 +1,12 @@
-import React, { lazy, Suspense } from "react";
+import React, { Suspense } from "react";
 import { Redirect, Route, Switch } from "react-router-dom";
 import { ContentRoute, LayoutSplashScreen } from "../_metronic/layout";
 import { DashboardPage } from "./pages/DashboardPage";
+import UserPage from "./pages/UserPage";
 
-const ECommercePage = lazy(() =>
-  import("./modules/ECommerce/pages/eCommercePage")
-);
+// const ECommercePage = lazy(() =>
+//   import("./modules/ECommerce/pages/eCommercePage")
+// );
 
 export default function BasePage() {
   // useEffect(() => {
@@ -21,7 +22,8 @@ export default function BasePage() {
           <Redirect exact from="/" to="/dashboard" />
         }
         <ContentRoute path="/dashboard" component={DashboardPage} />
-        <Route path="/e-commerce" component={ECommercePage} />
+        <Route path="/users" component={UserPage} />
+        {/* <Route path="/e-commerce" component={ECommercePage} /> */}
         <Redirect to="error/error-v1" />
       </Switch>
     </Suspense>
